requirements.txt
amqp==5.3.1 ; python_version >= "3.12" and python_version < "4.0"
apscheduler==3.11.1 ; python_version >= "3.12" and python_version < "4.0"
asgiref==3.10.0 ; python_version >= "3.12" and python_version < "4.0"
asttokens==3.0.1 ; python_version >= "3.12" and python_version < "4.0"
billiard==4.2.4 ; python_version >= "3.12" and python_version < "4.0"
black==25.11.0 ; python_version >= "3.12" and python_version < "4.0"
celery-types==0.23.0 ; python_version >= "3.12" and python_version < "4.0"
celery==5.6.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.11.12 ; python_version >= "3.12" and python_version < "4.0"
cfgv==3.5.0 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.12" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
click-plugins==1.1.1.2 ; python_version >= "3.12" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.3.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
coverage==7.13.0 ; python_version >= "3.12" and python_version < "4.0"
coverage[toml]==7.13.0 ; python_version >= "3.12" and python_version < "4.0"
crispy-bootstrap5==2025.6 ; python_version >= "3.12" and python_version < "4.0"
cron-descriptor==2.0.6 ; python_version >= "3.12" and python_version < "4.0"
decorator==5.2.1 ; python_version >= "3.12" and python_version < "4.0"
distlib==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
django-apscheduler==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
django-celery-beat==2.8.1 ; python_version >= "3.12" and python_version < "4.0"
django-cors-headers==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
django-crispy-forms==2.5 ; python_version >= "3.12" and python_version < "4.0"
django-filter==25.2 ; python_version >= "3.12" and python_version < "4.0"
django-timezone-field==7.2.1 ; python_version >= "3.12" and python_version < "4.0"
django==5.2.8 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework-simplejwt==5.5.1 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.16.1 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.8.0 ; python_version >= "3.12" and python_version < "4.0"
drf-yasg==1.21.11 ; python_version >= "3.12" and python_version < "4.0"
eventlet==0.40.4 ; python_version >= "3.12" and python_version < "4.0"
exceptiongroup==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
executing==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
filelock==3.20.3 ; python_version >= "3.12" and python_version < "4.0"
flake8==7.3.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==25.0.1 ; python_version >= "3.12" and python_version < "4.0"
identify==2.6.16 ; python_version >= "3.12" and python_version < "4.0"
idna==3.11 ; python_version >= "3.12" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
ipython-pygments-lexers==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
ipython==9.7.0 ; python_version >= "3.12" and python_version < "4.0"
isort==7.0.0 ; python_version >= "3.12" and python_version < "4.0"
jedi==0.19.2 ; python_version >= "3.12" and python_version < "4.0"
joblib==1.5.3 ; python_version >= "3.12" and python_version < "4.0"
kombu==5.6.1 ; python_version >= "3.12" and python_version < "4.0"
matplotlib-inline==0.2.1 ; python_version >= "3.12" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
nodeenv==1.10.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
parso==0.8.5 ; python_version >= "3.12" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
pexpect==4.9.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "emscripten"
pillow==12.0.0 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.5.0 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
poetry-core==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
pre-commit==4.5.1 ; python_version >= "3.12" and python_version < "4.0"
prompt-toolkit==3.0.52 ; python_version >= "3.12" and python_version < "4.0"
psycopg-binary==3.3.2 ; implementation_name != "pypy" and python_version >= "3.12" and python_version < "4.0"
psycopg[binary]==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "emscripten"
pure-eval==0.2.3 ; python_version >= "3.12" and python_version < "4.0"
pycodestyle==2.14.0 ; python_version >= "3.12" and python_version < "4.0"
pyflakes==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pytest-cov==7.0.0 ; python_version >= "3.12" and python_version < "4.0"
pytest-django==4.11.1 ; python_version >= "3.12" and python_version < "4.0"
pytest-mock==3.15.1 ; python_version >= "3.12" and python_version < "4.0"
pytest==9.0.2 ; python_version >= "3.12" and python_version < "4.0"
python-crontab==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.2.1 ; python_version >= "3.12" and python_version < "4.0"
pytokens==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.12" and python_version < "4.0"
redis==7.1.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.12" and python_version < "4.0"
scikit-learn==1.8.0 ; python_version >= "3.12" and python_version < "4.0"
scipy==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
setuptools==80.9.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.12" and python_version < "4.0"
stack-data==0.6.3 ; python_version >= "3.12" and python_version < "4.0"
threadpoolctl==3.6.0 ; python_version >= "3.12" and python_version < "4.0"
traitlets==5.14.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0"
tzlocal==5.3.1 ; python_version >= "3.12" and python_version < "4.0"
uritemplate==4.2.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.6.2 ; python_version >= "3.12" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.12" and python_version < "4.0"
virtualenv==20.36.1 ; python_version >= "3.12" and python_version < "4.0"
wcwidth==0.2.14 ; python_version >= "3.12" and python_version < "4.0"
